+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 **FlexLib**的所有版本的变更日志都将会在这里记录.
 
 ---
+1. 增加bundleForRes接口，bundleForImage默认使用bundleForRes的返回值
+2. 增加flexBasis属性
+3. iOS11以下添加状态栏高度
+
+
 ## 1.9.4
 1. 修复FlexModalView连续快速隐藏/显示切换的时候导致不显示的问题